move the `CompilationTarget` struct over into `target.rs`, cleanup usages, and change `crate::target::tuple()` to return a `CompilationTarget` instead of a String.
