+++ plrust/src/target.rs
@@ -4,8 +4,12 @@
     so a normal build-and-test cycle may create artifacts for multiple targets.
 */
 
+use once_cell::sync::Lazy;
 use std::env;
-use std::ffi::OsString;
+use std::ffi::{OsStr, OsString};
+use std::fmt::{Display, Formatter};
+use std::ops::Deref;
+use std::path::Path;
 
 mod host {
     use std::env::consts::ARCH;
@@ -54,24 +58,73 @@ pub(crate) enum TargetErr {
     InvalidSpec(OsString),
 }
 
-pub(crate) fn tuple() -> Result<String, TargetErr> {
-    match env::var("PLRUST_TARGET") {
-        Ok(v) => Ok(v),
-        Err(env::VarError::NotPresent) => {
-            cfg_if::cfg_if! {
-                if #[cfg(all(feature = "target_postgrestd",
-                    any(target_arch = "x86_64", target_arch = "aarch64"),
-                    target_os = "linux",
-                    target_env = "gnu"))]
-                {
-                    Ok(host::target_tuple())
-                } else if #[cfg(feature = "target_postgrestd")] {
-                    Err(TargetErr::Unsupported)
-                } else {
-                    Ok(host::target_tuple())
+#[derive(Debug, Clone, PartialOrd, PartialEq, Hash, Ord, Eq)]
+#[repr(transparent)]
+pub(crate) struct CompilationTarget(String);
+impl Deref for CompilationTarget {
+    type Target = str;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+impl From<&str> for CompilationTarget {
+    fn from(s: &str) -> Self {
+        CompilationTarget(s.into())
+    }
+}
+impl From<&String> for CompilationTarget {
+    fn from(s: &String) -> Self {
+        CompilationTarget(s.clone())
+    }
+}
+impl From<String> for CompilationTarget {
+    fn from(s: String) -> Self {
+        CompilationTarget(s)
+    }
+}
+impl Display for CompilationTarget {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.write_str(&self.0)
+    }
+}
+impl AsRef<Path> for CompilationTarget {
+    fn as_ref(&self) -> &Path {
+        Path::new(&self.0)
+    }
+}
+impl AsRef<OsStr> for CompilationTarget {
+    fn as_ref(&self) -> &OsStr {
+        OsStr::new(&self.0)
+    }
+}
+impl CompilationTarget {
+    pub fn as_str(&self) -> &str {
+        &self
+    }
+}
+
+pub(crate) fn tuple() -> Result<&'static CompilationTarget, &'static TargetErr> {
+    pub(crate) static TARGET_TUPLE: Lazy<Result<CompilationTarget, TargetErr>> =
+        Lazy::new(|| match env::var("PLRUST_TARGET") {
+            Ok(v) => Ok(v.into()),
+            Err(env::VarError::NotPresent) => {
+                cfg_if::cfg_if! {
+                    if #[cfg(all(feature = "target_postgrestd",
+                        any(target_arch = "x86_64", target_arch = "aarch64"),
+                        target_os = "linux",
+                        target_env = "gnu"))]
+                    {
+                        Ok(host::target_tuple().into())
+                    } else if #[cfg(feature = "target_postgrestd")] {
+                        Err(TargetErr::Unsupported)
+                    } else {
+                        Ok(host::target_tuple().into())
+                    }
                 }
             }
-        }
-        Err(env::VarError::NotUnicode(s)) => Err(TargetErr::InvalidSpec(s)),
-    }
+            Err(env::VarError::NotUnicode(s)) => Err(TargetErr::InvalidSpec(s)),
+        });
+
+    TARGET_TUPLE.as_ref()
 }
